+++ reductstore/src/api/bucket/get.rs
@@ -1,11 +1,11 @@
 // Copyright 2023 ReductStore
 // Licensed under the Business Source License 1.1
 
+use crate::api::bucket::FullBucketInfoAxum;
+use crate::api::middleware::check_permissions;
+use crate::api::Componentes;
+use crate::api::HttpError;
 use crate::auth::policy::AuthenticatedPolicy;
-use crate::http_frontend::bucket_api::FullBucketInfoAxum;
-use crate::http_frontend::middleware::check_permissions;
-use crate::http_frontend::Componentes;
-use crate::http_frontend::HttpError;
 use axum::extract::{Path, State};
 use axum::headers::HeaderMap;
 use std::sync::Arc;
@@ -30,9 +30,9 @@ pub async fn get_bucket(
 mod tests {
     use super::*;
 
-    use crate::http_frontend::Componentes;
+    use crate::api::Componentes;
 
-    use crate::http_frontend::tests::{components, headers};
+    use crate::api::tests::{components, headers};
 
     use rstest::rstest;
