+++ reductstore/src/api/bucket/head.rs
@@ -1,9 +1,9 @@
 // Copyright 2023 ReductStore
 // Licensed under the Business Source License 1.1
 
+use crate::api::middleware::check_permissions;
+use crate::api::{Componentes, HttpError};
 use crate::auth::policy::AuthenticatedPolicy;
-use crate::http_frontend::middleware::check_permissions;
-use crate::http_frontend::{Componentes, HttpError};
 
 use axum::extract::{Path, State};
 use axum::headers::HeaderMap;
@@ -24,9 +24,9 @@ pub async fn head_bucket(
 mod tests {
     use super::*;
 
-    use crate::http_frontend::Componentes;
+    use crate::api::Componentes;
 
-    use crate::http_frontend::tests::{components, headers};
+    use crate::api::tests::{components, headers};
 
     use rstest::rstest;
