+++ reductstore/src/api/bucket/update.rs
@@ -1,49 +1,42 @@
 // Copyright 2023 ReductStore
 // Licensed under the Business Source License 1.1
 
+use crate::api::bucket::BucketSettingsAxum;
+use crate::api::middleware::check_permissions;
+use crate::api::{Componentes, HttpError};
 use crate::auth::policy::FullAccessPolicy;
-use crate::http_frontend::bucket_api::BucketSettingsAxum;
-use crate::http_frontend::middleware::check_permissions;
-use crate::http_frontend::{Componentes, HttpError};
-
 use axum::extract::{Path, State};
 use axum::headers::HeaderMap;
 use std::sync::Arc;
 
-// POST /b/:bucket_name
-pub async fn create_bucket(
+// PUT /b/:bucket_name
+pub async fn update_bucket(
     State(components): State<Arc<Componentes>>,
     Path(bucket_name): Path<String>,
     headers: HeaderMap,
     settings: BucketSettingsAxum,
 ) -> Result<(), HttpError> {
     check_permissions(&components, headers, FullAccessPolicy {}).await?;
-    components
-        .storage
-        .write()
-        .await
-        .create_bucket(&bucket_name, settings.into())?;
-    Ok(())
+    let mut storage = components.storage.write().await;
+
+    Ok(storage
+        .get_mut_bucket(&bucket_name)?
+        .set_settings(settings.into())?)
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    use crate::http_frontend::tests::{components, headers};
-    use crate::http_frontend::Componentes;
-
-    use rstest::rstest;
-
+    use crate::api::tests::{components, headers};
     use reduct_base::error::ErrorCode;
-    use std::sync::Arc;
+    use rstest::rstest;
 
     #[rstest]
     #[tokio::test]
-    async fn test_create_bucket(components: Arc<Componentes>, headers: HeaderMap) {
-        create_bucket(
+    async fn test_update_bucket(components: Arc<Componentes>, headers: HeaderMap) {
+        update_bucket(
             State(components),
-            Path("bucket-3".to_string()),
+            Path("bucket-1".to_string()),
             headers,
             BucketSettingsAxum::default(),
         )
@@ -53,10 +46,10 @@ mod tests {
 
     #[rstest]
     #[tokio::test]
-    async fn test_create_bucket_already_exists(components: Arc<Componentes>, headers: HeaderMap) {
-        let err = create_bucket(
+    async fn test_update_bucket_not_found(components: Arc<Componentes>, headers: HeaderMap) {
+        let err = update_bucket(
             State(components),
-            Path("bucket-1".to_string()),
+            Path("not-found".to_string()),
             headers,
             BucketSettingsAxum::default(),
         )
@@ -65,7 +58,7 @@ mod tests {
         .unwrap();
         assert_eq!(
             err,
-            HttpError::new(ErrorCode::Conflict, "Bucket 'bucket-1' already exists",)
+            HttpError::new(ErrorCode::NotFound, "Bucket 'not-found' is not found",)
         )
     }
 }
