+++ reductstore/src/api/entry/read_single.rs
@@ -1,11 +1,11 @@
 // Copyright 2023 ReductStore
 // Licensed under the Business Source License 1.1
 
+use crate::api::entry::{check_and_extract_ts_or_query_id, MethodExtractor};
+use crate::api::middleware::check_permissions;
+use crate::api::Componentes;
+use crate::api::HttpError;
 use crate::auth::policy::ReadAccessPolicy;
-use crate::http_frontend::entry_api::{check_and_extract_ts_or_query_id, MethodExtractor};
-use crate::http_frontend::middleware::check_permissions;
-use crate::http_frontend::Componentes;
-use crate::http_frontend::HttpError;
 use crate::storage::bucket::Bucket;
 
 use crate::storage::reader::RecordReader;
@@ -137,7 +137,7 @@ mod tests {
 
     use axum::body::HttpBody;
 
-    use crate::http_frontend::tests::{components, headers, path_to_entry_1};
+    use crate::api::tests::{components, headers, path_to_entry_1};
     use crate::storage::query::base::QueryOptions;
     use reduct_base::error::ErrorCode;
     use reduct_base::error::ErrorCode::NotFound;
