+++ reductstore/src/cfg.rs
@@ -1,11 +1,11 @@
 // Copyright 2023 ReductStore
 // Licensed under the Business Source License 1.1
 
+use crate::api::Componentes;
 use crate::asset::asset_manager::ZipAssetManager;
 use crate::auth::token_auth::TokenAuthorization;
 use crate::auth::token_repository::{create_token_repository, ManageTokens};
 use crate::core::env::{Env, GetEnv};
-use crate::http_frontend::Componentes;
 
 use crate::storage::storage::Storage;
 use bytesize::ByteSize;
