+++ reductstore/src/lib.rs
@@ -1,8 +1,8 @@
 // Copyright 2023 ReductStore
 // Licensed under the Business Source License 1.1
+pub mod api;
 pub mod asset;
 pub mod auth;
 pub mod cfg;
 pub mod core;
-pub mod http_frontend;
 pub mod storage;
